std::span sucks why'd I start this

+++ Capsule.h
@@ -44,10 +44,10 @@ public:
 	glm::vec3 ClosestPoint(const glm::vec3& other) const;
 
 	void GenerateMesh(Buffer<ArrayBuffer>& verts, Buffer<ElementArray>& indicies,
-		const std::size_t latitudeSlices = 18, const std::size_t longitudeSlices = 18) const;
+		const std::uint8_t latitudeSlices = 18, const std::uint8_t longitudeSlices = 18) const;
 
 	static void GenerateMesh(Buffer<ArrayBuffer>& verts, Buffer<ElementArray>& indicies, float radius, float distance,
-		const std::size_t latitudeSlices = 18, const std::size_t longitudeSlices = 18);
+		const std::uint8_t latitudeSlices = 18, const std::uint8_t longitudeSlices = 18);
 };
 
 constexpr Capsule::Capsule(const Capsule& other) noexcept : line(other.line), radius(other.radius) {}
